Remove Py37+ test gating on async generators
